Unity 2023.1.13f1 C# reference source code

+++ Modules/TextCoreTextEngine/Managed/FontAsset.cs
@@ -1224,7 +1224,7 @@ namespace UnityEngine.TextCore.Text
         /// <returns></returns>
         public bool HasCharacter(char character, bool searchFallbacks = false, bool tryAddCharacter = false)
         {
-            return HasCharacter(character, searchFallbacks, tryAddCharacter);
+            return HasCharacter((uint)character, searchFallbacks, tryAddCharacter);
         }
 
         /// <summary>
